Added properties in Table. Created function Pay in Business class.
Almost finished with Pay function, but remains logic for when guest does not have enough money - SendGuestToWashDishes.

+++ TheRestaurant/Persons/Guest.cs
@@ -10,7 +10,7 @@ namespace TheRestaurant.Persons
 {
     internal class Guest : Person
     {
-        public int Money { get; set; }
+        public double Money { get; set; }
         public int Satisfied { get; set; }
         public bool IsVegetarian { get; set; }
         public Food MyMeal { get; set; }
